feat: desenvolvimento crud + requisições http para entidades Professor

+++ src/main/java/com/altf7/sei/repository/ProfessorRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface ProfessorRepository extends JpaRepository<Professor, Long> {
+public interface ProfessorRepository extends JpaRepository<Professor, Integer> {
     Optional<Professor> findByCpf(String cpf);
 }
